feat: 实现 ModLoader 热重载和模块加载功能
新增功能：

ModLoader 热重载机制
   - 实现动态加载和卸载 Mod 的能力
   - 支持运行时重新加载修改的 Mod
   - 提供热重载回调接口，支持状态通知

测试通过：
- Module 动态加载和初始化
- Mod 的热重载功能

+++ include/patchlib/type.h
@@ -47,8 +47,8 @@ extern "C" {
 typedef void *patch_handle_t;
 #define PATCH_NULL NULL
 #else
-typedef int32_t patch_handle_t;
-#define PATCH_NULL (-1)
+typedef intptr_t patch_handle_t;
+#define PATCH_NULL ((intptr_t)-1)
 #endif
 
 // 基本类型枚举
